cache ffmpeg

+++ pkgs/ffmpeg/fetch.sh
@@ -1,8 +1,12 @@
 #!/bin/sh
 rm -rf pkg src
-printf "Downloading source...\n"
-curl -L -# -o src.tar.xz https://ffmpeg.org/releases/ffmpeg-8.0.tar.xz
+ver='8.0'
+[ -z "$_DLCACHE" ] && _DLCACHE=/tmp
+if [ ! -f "$_DLCACHE/ffmpeg-$ver.tar.xz" ] ||
+    [ "$(sha256sum "$_DLCACHE/ffmpeg-$ver.tar.xz" | awk '{print $1}')" != "b2751fccb6cc4c77708113cd78b561059b6fa904b24162fa0be2d60273d27b8e" ]; then
+    printf "Downloading source...\n"
+    curl -L -# -o "$_DLCACHE/ffmpeg-$ver.tar.xz" "https://ffmpeg.org/releases/ffmpeg-$ver.tar.xz" || exit 1
+fi
 printf "Unpacking source...\n"
-tar -xf src.tar.xz
-rm src.tar.xz
+tar -xf "$_DLCACHE/ffmpeg-$ver.tar.xz"
 mv ffmpeg-* src
